+++ crates/wdl-grammar/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+#### Removed
+
+* Removed the `SyntaxExt` trait ([#482](https://github.com/stjude-rust-labs/wdl/pull/482))
+
 ## 0.13.0 - 05-02-2025
 
 #### Dependencies